Aggregate Treatment Effect Parameters Object

Objects of this class hold results on aggregated unit-time average treatment effects

Usage

AGGITEobj(
  overall.att = NULL,
  overall.se = NULL,
  overall.lci = NULL,
  overall.uci = NULL,
  type = "simple",
  type2 = NULL,
  egt = NULL,
  egt2 = NULL,
  att.egt = NULL,
  se.egt = NULL,
  lci.egt = NULL,
  uci.egt = NULL,
  crit.val.egt = NULL,
  inf.function = NULL,
  min_e = NULL,
  max_e = NULL,
  balance_e = NULL,
  min_agg = NULL,
  call = NULL,
  DIDparams = NULL
)

Arguments

overall.att

The estimated overall ATT

overall.se

Standard error for overall ATT

overall.lci

The lower confidence interval of the overall ATT

overall.uci

The upper confidence interval of the overall ATT

type

Which type of aggregated treatment effect parameter to compute. One option is "simple" (this just computes a weighted average of all unit-time average treatment effects). Other options are "dynamic" (this computes average effects across different lengths of exposure to the treatment and is similar to an "event study"; here the overall effect averages the effect of the treatment across all positive lengths of exposure); "group" (this is the default option and computes average treatment effects across different groups; here the overall effect averages the effect across different groups); and "calendar" (this computes average treatment effects across different time periods; here the overall effect averages the effect across each time period).

type2

the second type of aggregator

egt

Holds the length of exposure (for dynamic effects), the group (for selective treatment timing), the unit (for selective treatment), cohort (for cohort level effects), or the time period (for calendar time effects)

egt2

a second aggregation type to hold a secondary object, NULL if no such object

att.egt

The ATT specific to egt

se.egt

The standard error specific to egt from the bootstrap

lci.egt

The lower confidence interval from the bootstrap

uci.egt

The upper confidence interval from the bootstrap

crit.val.egt

A critical value for computing uniform confidence bands for dynamic effects, selective treatment timing, or time period effects.

inf.function

The bootstrap draws specific to the egt

min_e

For event studies, this is the smallest event time to compute dynamic effects for. By default, min_e = -Inf so that effects at all lengths of exposure are computed.

max_e

For event studies, this is the largest event time to compute dynamic effects for. By default, max_e = Inf so that effects at all lengths of exposure are computed.

balance_e

If set (and if one computes dynamic effects), it balances the sample with respect to event time. For example, if balance.e=2, aggite will drop groups that are not exposed to treatment for at least three periods. (the initial period when e=0 as well as the next two periods when e=1 and the e=2). This ensures that the composition of groups does not change when event time changes.

min_agg

the minimum number of unit-time effects required for aggregation. Default is 2.

call

The function call to aggite2

DIDparams

A DIDparams_i object

Value

an AGGITEobj

Examples

# A helper function for [aggite()]. See examples in documentation in that function.