MP_i

Multi-period objects that hold results for unit-time average treatment effects

Usage

MP_i(
  id,
  group,
  t,
  att,
  V_analytical,
  se,
  c,
  lci,
  uci,
  inffunc,
  n = NULL,
  aggite = NULL,
  alp = 0.05,
  ipwqual = NULL,
  attcalc = NULL,
  baseline = NULL,
  baset = NULL,
  outcome = NULL,
  count = NULL,
  DIDparams = NULL
)

Arguments

id

which unit the unit-time treatment effects are for

group

which group (defined by period first treated) an unit-time average treatment effect is for

t

which time period a group-time average treatment effect is for

att

the estimate for unit id and time period t

V_analytical

Analytical estimator for the asymptotic variance-covariance matrix for unit-time average treatment effects

se

standard errors for unit-time average treatment effects. If bootstrap is set to TRUE, this provides bootstrap-based se.

c

always NULL

lci

lower confidence interval for att

uci

upper confidence interval for att

inffunc

the influence function for estimating group-time average treatment effects

n

the number of unique cross-sectional units (unique values of idname)

aggite

an aggregate treatment effects object

alp

the significance level, default is 0.05

ipwqual

the maximum propensity score

attcalc

similar to att but does not remove estimates when there are propensity score problems

baseline

the baseline of the unit

baset

the baseline time of the unit

outcome

the outcome of the unit at time t

count

takes a count of the number of units in the component estimate

DIDparams

a DIDparams_i object. A way to optionally return the parameters of the call to att_it().

Value

MP object

Examples

# Helper function for [att_it()]. See documentation of that function for an example.