Unit-Time Average Treatment Effects

Computes the 2x2 DID estimate at the unit-time level and allows for unit-level covariate matching in a staggered treatment setting. The original package by Brantly Callaway and Pedro Santa'Anna (did) only allows for group-time ATTs where units treated in the same period is in a single group. Unlike the related did package, there is no provision for repeated cross-sections.

Usage

att_it(
  yname,
  tname,
  idname,
  gname,
  cohortnames = NULL,
  xformla = NULL,
  data,
  panel = TRUE,
  control_group = c("nevertreated", "notyettreated"),
  anticipation = 0,
  weightsname = NULL,
  alp = 0.05,
  bstrap = TRUE,
  cband = TRUE,
  biters = 1000,
  clustervars = NULL,
  est_method = "dr",
  overlap = "trim",
  base_period = "varying",
  print_details = FALSE,
  pl = FALSE,
  cores = 1
)

Arguments

yname

the name of the outcome variable in the data

tname

the name of the column containing the time periods in data. Must be numeric

idname

the individual (cross-sectional unit) id name in data. Cannot be missing and must be numeric.

gname

the name of the variable in data that contains the first period when a particular observation is treated. This should be a positive number for all observations in treated groups. It defines which "treatment-group" a unit belongs to. A zero (0) value is considered to be a unit that is never treated. A unit treated after the last period in tname is also considered never treated.

cohortnames

the names of additional aggregation variables in data.

xformla

a formula for the covariates to include in the model. It should be of the form ~ X1 + X2. An intercept is automatically included. Default is NULL which is equivalent to xformla=~1.

data

the name of the data.frame that contains the data.

panel

if panel is TRUE, the data is coerced to a balanced panel on the. Default is TRUE.

control_group

which units to use the control group. The default is "nevertreated" which sets the control group to be the group of units that never participate in the treatment. This group does not change across groups or time periods. The other option is to set group="notyettreated". In this case, the control group is set to the group of units that have not yet participated in the treatment in that time period. This includes all never treated units, but it includes additional units that eventually participate in the treatment, but have not participated yet.

anticipation

the number of time periods before participating in the treatment where units can anticipate participating in the treatment and therefore it can affect their untreated potential outcomes.

weightsname

the name of the column containing the sampling weights. If not set, all observations have same weight.

alp

the significance level, default is 0.05

bstrap

boolean for whether or not to compute standard errors using the multiplier bootstrap. If standard errors are clustered, then one must set bstrap=TRUE. Default is TRUE (in addition, cband is also by default TRUE indicating that uniform confidence bands will be returned. If bstrap is FALSE, then analytical standard errors are reported.

cband

boolean for whether or not to compute a uniform confidence band that covers all of the group-time average treatment effects with fixed probability 1-alp. In order to compute uniform confidence bands, bstrap must also be set to TRUE. The default is TRUE.

biters

the number of bootstrap iterations to use. The default is 1000, and this is only applicable if bstrap=TRUE.

clustervars

a vector of variables names to cluster on. At most, there can be two variables (otherwise will throw an error) and one of these must be the same as idname which allows for clustering at the individual level. By default, we cluster at individual level (when bstrap=TRUE)

est_method

the method to compute group-time average treatment effects. The default is "dr" which uses the doubly robust approach in the DRDID package. Other built-in methods include "ipw" for inverse probability weighting (Hajek type) and "reg" for first step regression estimators.

overlap

the treatment of units that violate overlap conditions when the est_method is "dr" or "ipw". The default, "trim", is to drop the unit but report the calculated ATT for further analysis. Overlap is violated if the maximum pscore exceeds 0.999 The other option, "retain" retains these units for inference.

base_period

whether to use a "varying" base period or a "universal" base period for placebo tests. A varying base period calculates a pseudo-ATT for every two consecutive pre-treatment periods. A universal base period fixes the base period to always be (g-anticipation-1). Either choice results in the same post-treatment estimates of ATT(g,t)'s.

print_details

whether or not to show details/progress of computations. Default is FALSE.

pl

whether or not to use parallel processing

cores

the number of cores to use for parallel processing

Value

An MP_i object containing all the results for unit-time treatment effects.

Examples

# run the function with the default settings and default data
attobject = att_it(yname = "y", tname = "time", gname = "treatg", idname ="unit", data = sim_data())