GDP and democratization data (1960-2010)

A dataset containing the gross domestic product per capita for 175 countries from 1960 to 2010. This is a subset of the dataset used in:

Usage

demgdp

Format

demgdp

A data frame with 6934 rows and 7 columns:

year

the year of the observation

country_name

the country name

wbcode

a unique identifier for a particular country

gdppercapitaconstant2000us

the GDP per capita in US dollars

dem

the indicator for whether a country is a democracy (1) or not (0)

YearFirstDemocracy

the year that the country first becomes a democracy since 1960. If already a democracy this is 1960, and if not a democracy by 2010, this is missing

breakdown

indicator for if the country has reverted to an autocracy since democratizing in 1960

...

Source

https://economics.mit.edu/people/faculty/daron-acemoglu/data-archive

Details

Acemoglu, D., Naidu, S., Restrepo, P., & Robinson, J. A. (2019). Democracy does cause growth. Journal of political economy, 127(1), 47-100.