Creates a simulated dataset with treatment dosages and a control group

Creates a simulated dataset with treatment dosages and a control group

Usage

sim_data(
  basetreat = seq(10, 100, 10),
  timetreat = rep(c(10, 15), 5),
  dosage = rep(c(1, 1), each = 5),
  tef = rep(10, 10),
  basecontrol = seq(5, 100, 5),
  cohort = rep(c(1, 1, 1, 1, 1), 6),
  pretreat = 5,
  posttreat = 20,
  ttr = 2,
  siget = 1,
  sigec = 1,
  fixedbase = FALSE
)

Arguments

basetreat

a numeric vector that specifies the pre-treatment outcomes

timetreat

a numeric vector of the same length as basetreat that specifies the treatment time (avoid negative or zero times)

dosage

a numeric vector of the same length as basetreat that specifies the treatment dosage received by a treatment unit. Zero for control units.

tef

a numeric vector of the same length as basetreat that specifies the treatment effect. The argument dosage multiplies the treatment effect by the number of doses.

basecontrol

a numeric vector that specifies the outcome level at the earliest treatment time

cohort

a vector of the length of basecontrol and basetreat combined that identifies the cohort of the unit

pretreat

the number of pre-treatment periods for the earliest treated unit

posttreat

the number of post-treatment periods for the earliest treated unit. If units are treated posttreat periods after the earliest treated unit, they will never be treated.

ttr

the time trend

siget

the standard error of the error term of the treatment group with default 1

sigec

the standard error of the error term of the control group with default 1

fixedbase

baseline observation is known and therefore is not perturbed.

Value

a dataframe with identifies for unit, time, treatment time, cohort, dosage, and the outcome

Examples

# A default dataframe with only one dosage level and one cohort
sim_data()
# A default dataframe with two different dosage levels
sim_data(dosage = rep(c(1,2),each=5))
# A default dataframe with 5 different cohorts
sim_data(cohort = rep(c(1,2,3,4,5),6))