An archived version of the DRDID's std_ipw_did_panel without the fastglm function

An archived version of the DRDID's std_ipw_did_panel without the fastglm function

Usage

std_ipw_did_panel_archive(
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  boot = FALSE,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = FALSE
)

Arguments

y1

The post-treatment vector

y0

The pre-treatment vector

D

The binary treatment vector

covariates

The matrix or vector of covariates to train weights

i.weights

Sample weights. Default is NULL.

boot

A logical as to whether to bootstrap. Default is FALSE.

boot.type

Type of bootstrap. Default is a weighted bootstrap.

nboot

Number of bootstraps

inffunc

A logical as to whether to report the influence function or not.

Value

Returns an object with calculated ATT, standard error, confidence intervals, and influence functions.

Examples

std_ipw_did_panel_archive(rep(1,20),rep(0,20),c(1,rep(0,19)),c(70,rep(50,9),rep(80,10)))
#> $ATT
#> [1] 0
#> 
#> $se
#> [1] 0
#> 
#> $uci
#> [1] 0
#> 
#> $lci
#> [1] 0
#> 
#> $boots
#> NULL
#> 
#> $att.inf.func
#> NULL
#> 
#> $call.param
#> std_ipw_did_panel_archive(y1 = rep(1, 20), y0 = rep(0, 20), D = c(1, 
#>     rep(0, 19)), covariates = c(70, rep(50, 9), rep(80, 10)))
#> 
#> $argu
#> $argu$panel
#> [1] TRUE
#> 
#> $argu$normalized
#> [1] TRUE
#> 
#> $argu$boot
#> [1] FALSE
#> 
#> $argu$boot.type
#> [1] "weighted"
#> 
#> $argu$nboot
#> NULL
#> 
#> $argu$type
#> [1] "ipw"
#> 
#>